requirements.txt
Flask==2.3.3
langchain-openai==0.2.2
openai==1.57.0
python-dotenv==1.0.1
flask-cors==4.0.0
# Let pip pick compatible langchain + langchain-core versions
langchain>=0.3.0
